[feat,resolve] 채팅 1:1 목록조회 오류 수정, 채팅 수정 삭제 기능 추가

+++ src/main/java/com/hunmin/domain/service/ChatRoomService.java
@@ -15,25 +15,25 @@ import com.hunmin.domain.repository.MemberRepository;
 import jakarta.annotation.Resource;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.redis.core.HashOperations;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-import java.util.*;
 import java.io.IOException;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
 @Log4j2
 public class ChatRoomService {
 
-    @Resource(name = "redisTemplate")
-    private HashOperations<String, String, ChatRoomRequestDTO> roomStorage;
+    private final HashOperations<String, String, Object> roomStorage;
     private final ObjectMapper objectMapper;
-
     private final ChatRoomRepository chatRoomRepository;
     private final MemberRepository memberRepository;
-
     private final NotificationService notificationService;
     private final SseEmitters sseEmitters;
 
@@ -43,40 +43,40 @@ public class ChatRoomService {
         return new ChatRoomDTO(chatRoom);
     }
 
-    // 나랑 관련된 모든 채팅방 조회
     public List<ChatRoomRequestDTO> findRoomByEmail(String email) {
         Member me = memberRepository.findByEmail(email);
-        List<ChatRoomRequestDTO> partnerNameAndChatRoom = roomStorage.values(me.getNickname());
-        log.info("partnerNameAndChatRoom={}", partnerNameAndChatRoom);
 
+        // Redis에서 해당 사용자의 모든 채팅방 정보를 가져옵니다.
+        List<Object> partnerNameAndChatRoom = roomStorage.values(me.getNickname());
+        log.info("partnerNameAndChatRoom={}", partnerNameAndChatRoom.toString());
+
+        Set<Long> chatRoomIds = new HashSet<>();
         List<ChatRoomRequestDTO> chatRoomRequestDTOList = new ArrayList<>();
+        log.info("chatRoomRequestDTOList1={}", chatRoomRequestDTOList);
 
-        Set<Long> chatRoomIdInSet = new HashSet<>();
-        for (Object rawChatRoom : partnerNameAndChatRoom) {
-            ChatRoomRequestDTO chatRoomRequestDTO;
-            chatRoomRequestDTO = objectMapper.convertValue(rawChatRoom, ChatRoomRequestDTO.class);
-            if (chatRoomRequestDTO != null) {
-                chatRoomIdInSet.add(chatRoomRequestDTO.getChatRoomId());
-                chatRoomRequestDTOList.add(chatRoomRequestDTO);
-                log.info("chatRoomIdInSet={}", chatRoomIdInSet);
-                log.info("chatRoomRequestDTOList={}", chatRoomRequestDTOList);
-            }
+        for (Object chatRoomRequestDTO : partnerNameAndChatRoom) {
+            log.info("chatRoomRequestDTO1={}", chatRoomRequestDTO.toString());
+            ChatRoomRequestDTO chatRoomRequest = objectMapper.convertValue(chatRoomRequestDTO, ChatRoomRequestDTO.class);
+            chatRoomIds.add(chatRoomRequest.getChatRoomId());
+            chatRoomRequestDTOList.add(chatRoomRequest);
         }
 
-        Set<String> partnerNames = roomStorage.keys(me.getNickname());
-        for (String partnerName : partnerNames) {
-            log.info("partnerName={}", partnerName);
-            Object rawChatRoom = roomStorage.get(partnerName, me.getNickname());
-            ChatRoomRequestDTO chatRoomRequestDTO = objectMapper.convertValue(rawChatRoom, ChatRoomRequestDTO.class);
+        List<Member> allMembers = memberRepository.findAll();
+        for (Member memberIndex : allMembers) {
+            log.info("member number={}", memberIndex.toString());
+            Object rawChatRoom = roomStorage.get(memberIndex.getNickname(), me.getNickname());
+            ChatRoomRequestDTO chatRoomRequestDTO=objectMapper.convertValue(rawChatRoom, ChatRoomRequestDTO.class);
             if (chatRoomRequestDTO != null) {
-                log.info("chatRoomRequestDTO={}", chatRoomRequestDTO);
-                if (!chatRoomIdInSet.contains(chatRoomRequestDTO.getChatRoomId())) {
+                log.info("chatRoomRequestDTO2={}", chatRoomRequestDTO.toString());
+                if (!chatRoomIds.contains(chatRoomRequestDTO.getChatRoomId())) {
+                    chatRoomIds.add(chatRoomRequestDTO.getChatRoomId());
                     chatRoomRequestDTOList.add(chatRoomRequestDTO);
                 }
             }
         }
         return chatRoomRequestDTOList;
     }
+
 
     // 채팅방 생성 : 이름으로 상대방 검색 후 채팅방 개설  -> 새로운
     public ChatRoomRequestDTO createChatRoomByNickName(String partnerName, String myEmail) {
@@ -88,7 +88,6 @@ public class ChatRoomService {
 
         Member me = memberRepository.findByEmail(myEmail);
         log.info("me ={}", me);
-        ChatRoomRequestDTO chatRoomFromStorage = roomStorage.get(me.getNickname(), partnerName);
 
         String myNickname = me.getNickname();
         // 이미 존재하는지 확인
@@ -98,13 +97,14 @@ public class ChatRoomService {
         }
         ChatRoom chatRoom = ChatRoom.builder().member(me).build();
         ChatRoom SavedchatRoom = chatRoomRepository.save(chatRoom);
-        roomStorage.put(me.getNickname(), partnerName, ChatRoomRequestDTO.builder()
+        ChatRoomRequestDTO chatRoomRequestDTO = ChatRoomRequestDTO.builder()
                 .chatRoomId(SavedchatRoom.getChatRoomId())
                 .memberId(me.getMemberId())
                 .nickName(me.getNickname())
                 .partnerName(partnerName)
                 .createdAt(SavedchatRoom.getCreatedAt())
-                .build());
+                .build();
+        roomStorage.put(me.getNickname(), partnerName, chatRoomRequestDTO);
         log.info("roomStorage ={}", roomStorage);
 
         Member partner = memberRepository.findByNickname(partnerName).orElseThrow(MemberException.NOT_FOUND::get);
@@ -145,14 +145,22 @@ public class ChatRoomService {
     // 채팅방 삭제
     public Boolean deleteChatRoom(Long chatRoomId, String partnerName, String meEmail) {
 
-        Member me = memberRepository.findByEmail(meEmail);
+        Optional<Member> partner = memberRepository.findByNickname(partnerName);
+        if (partner.isEmpty()) {
+            return false;
+        }
 
+        Member me = memberRepository.findByEmail(meEmail);
         ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
                 .orElseThrow(ChatRoomException.NOT_FOUND::get);
 
-        roomStorage.delete(me.getNickname(), partnerName);
-
+        if (roomStorage.get(me.getNickname(), partnerName) != null) {
+            roomStorage.delete(me.getNickname(), partnerName);
+        } else if (roomStorage.get(partnerName, me.getNickname()) != null) {
+            roomStorage.delete(partnerName, me.getNickname());
+        } else return false;
         chatRoomRepository.delete(chatRoom);
+
         return true;
     }
 }
